tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/142.html

<!DOCTYPE html>
<html><head>
	<title>scheduler: SVG inline script that document.writes external script </title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>
  var t = async_test(undefined, {timeout:3000});
</script>
<svg>
<script>
  t.step(function() {
    log('inline script #1');
    document.write("<" + "script xlink:href='scripts/include-1.js'><" + "/script>");
    log('end inline script #1');
  });
</script>
<script>t.step(function() {log("inline script #2")});</script>
</svg>
<script>
onload = t.step_func(function() {
  assert_array_equals(eventOrder, ["inline script #1", "end inline script #1", "external script #1", "inline script #2"]);
  t.done();
});
</script>
